feat(tv): 粗对齐 Kotlin TV 页面壳样式

+++ re-android/core-data/src/main/java/org/moontechlab/selene/tv/core/data/model/TvVideoCard.kt
@@ -28,7 +28,7 @@ data class TvVideoCard(
     val episodeIndex: Int = 0,
     val playTime: Int = 0,
     val totalTime: Int = 0,
-    val saveTime: Int = 0,
+    val saveTime: Long = 0L,
     val searchTitle: String = "",
     val origin: String = "",
 )
